Phong upcode add CRUD cơ bản và đổi sang code first

+++ code/Garage-Finder-Backend/GFData/Models/Entity/Service.cs
@@ -9,8 +9,8 @@ namespace GFData.Models.Entity
         public int GarageID { get; set; }
         public string NameService { get; set; }
         public int CategoryID { get; set; }
-        public float Cost { get; set; }
+        public double Cost { get; set; }
         public string Note { get; set; }
-        public virtual Category? Category { get; set; }
+        public virtual Category? Category { get; set; } 
     }
 }
